refactor tornado_handlers: split up into separate files
It got too large
