+++ run_pylint.sh
@@ -8,7 +8,7 @@ pylint_exec=$(which pylint 2>/dev/null)
 set -e
 
 export PYTHONPATH=plot_app
-python $pylint_exec send_email.py tornado_handlers.py serve.py \
-	multipart_streamer.py plot_app/*.py
+python $pylint_exec tornado_handlers/*.py serve.py \
+	plot_app/*.py
 
 exit 0
